requirements.txt
pandas
mysql-connector-python
snowflake-connector-python
python-dotenv
